+++ app/src/main/java/uk/co/oliverdelange/wcr_android_kt/ui/view/PaintableTouchImageView.kt
@@ -111,7 +111,7 @@ class PaintableTouchImageView(c: Context, att: AttributeSet) : TouchImageView(c,
 }
 
 class PathCapture : Path() {
-    val capture = mutableSetOf<Pair<Float, Float>>()
+    val capture = mutableSetOf<Pair<Int, Int>>()
 
     override fun reset() {
         super.reset()
@@ -120,16 +120,16 @@ class PathCapture : Path() {
 
     override fun moveTo(x: Float, y: Float) {
         super.moveTo(x, y)
-        capture.add(Pair(x, y))
+        capture.add(Pair(x.toInt(), y.toInt()))
     }
 
     override fun quadTo(x1: Float, y1: Float, x2: Float, y2: Float) {
         super.quadTo(x1, y1, x2, y2)
-        capture.add(Pair(x2, y2))
+        capture.add(Pair(x2.toInt(), y2.toInt()))
     }
 
     override fun lineTo(x: Float, y: Float) {
         super.lineTo(x, y)
-        capture.add(Pair(x, y))
+        capture.add(Pair(x.toInt(), y.toInt()))
     }
 }
